♻️ apply clean architecture design
