+++ app/container.py
@@ -3,9 +3,9 @@
 import injector
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
+from app.domain.repositories import IUnitOfWork
 from app.infrastructure.orm_registry import init_orm_mappings
 from app.infrastructure.unit_of_work import SQLAlchemyUnitOfWork
-from app.repository import IUnitOfWork
 
 
 class DatabaseModule(injector.Module):
