+++ app/domain/repositories/__init__.py
@@ -1,0 +1,17 @@
+"""Repository interfaces for domain layer."""
+
+from app.domain.repositories.interfaces import (
+    IRepository,
+    IRepositoryWithId,
+    IUnitOfWork,
+    RepositoryError,
+    RepositoryErrorType,
+)
+
+__all__ = [
+    "IRepository",
+    "IRepositoryWithId",
+    "IUnitOfWork",
+    "RepositoryError",
+    "RepositoryErrorType",
+]
