+++ tests/infrastructure/test_repositories.py
@@ -7,8 +7,8 @@ import pytest
 
 from app.core.result import Err, Ok
 from app.domain.aggregates.user import User
+from app.domain.repositories import IUnitOfWork
 from app.domain.value_objects import Email, UserId
-from app.repository import IUnitOfWork
 
 
 @pytest.mark.anyio
