+++ tests/infrastructure/test_team_repositories.py
@@ -7,8 +7,8 @@ import pytest
 
 from app.core.result import Err, Ok
 from app.domain.aggregates.team import Team
+from app.domain.repositories import IUnitOfWork
 from app.domain.value_objects import TeamId, TeamName
-from app.repository import IUnitOfWork
 
 
 @pytest.mark.anyio
