+++ tests/usecases/users/test_get_user.py
@@ -4,8 +4,8 @@ import pytest
 
 from app.core.result import Err, Ok
 from app.domain.aggregates.user import User
+from app.domain.repositories import IUnitOfWork
 from app.domain.value_objects import Email, UserId
-from app.repository import IUnitOfWork
 from app.usecases.result import ErrorType
 from app.usecases.users.get_user import GetUserHandler, GetUserQuery
